project/my_django_js_webapp/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
blinker==1.8.2
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
cron-descriptor==1.4.5
Django==4.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-timezone-field==7.0
djangorestframework==3.15.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
flower==2.0.1
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.1.1
humanize==4.10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
locust==2.31.7
MarkupSafe==2.1.5
msgpack==1.1.0
prometheus_client==0.21.0
prompt_toolkit==3.0.47
psutil==6.0.0
PyMySQL==1.1.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
XlsxWriter==3.2.0
zope.event==5.0
zope.interface==7.0.3
